% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{all_filenames}
\alias{all_filenames}
\title{all_filenames: function to get all file names from a directory of .dxf files}
\usage{
all_filenames(path)
}
\arguments{
\item{path}{path to the directory of .dxf files}
}
\value{
vector of filenames for all .dxf files in the specified directory
}
\description{
all_filenames: function to get all file names from a directory of .dxf files
}
\examples{
Usage example
all_filenames("~/path_to_files")
}
