% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{all_peak_areas_poly}
\alias{all_peak_areas_poly}
\title{all_peak_areas_poly: function to get all peak areas in an experiment}
\usage{
all_peak_areas_poly(start.vec, end.vec, time.vec, int.vec)
}
\arguments{
\item{start.vec}{numeric vector containing all peak start times}

\item{end.vec}{numeric vector containing all peak end times}

\item{time.vec}{numeric vector containing all the raw time.s data}

\item{int.vec}{numeric vector continaing all the raw intentisty data (i.e. v44.mV, etc,)}
}
\value{
a vector containing the areas of each peak in V*s
}
\description{
all_peak_areas_poly: function to get all peak areas in an experiment
}
\examples{
Usage example
all_peak_areas_poly(start.v1,end.v1,allt.s,allv44)
}
