% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{all_peak_areas_trap}
\alias{all_peak_areas_trap}
\title{all_peak_areas_trap: function to calculate all peak areas using the integration trapezoidal rule via trapz}
\usage{
all_peak_areas_trap(start.vec, end.vec, time.vec, int.vec)
}
\arguments{
\item{start.vec}{numeric vector of peak start times}

\item{end.vec}{numeric vector of peak end times}

\item{time.vec}{numeric vector of times from the raw data}

\item{int.vec}{numeric vector of intensities from the raw data}
}
\value{
a numeric vector containing all the peak areas in V*s
}
\description{
all_peak_areas_trap: function to calculate all peak areas using the integration trapezoidal rule via trapz
}
\examples{
Usage Example
all_peak_areas_trap<-all_peak_areas_trap(start.v1,end.v1,time.s,v44)
}
