% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{average_times}
\alias{average_times}
\title{average_times: function that finds the averages of each peak time for vendor data (called in find_expected_times())}
\usage{
average_times(peakTimes.list, nPeaks)
}
\arguments{
\item{peakTimes.list}{list of peak numbers and times from vendor table (output from peak_vend_times_all())}

\item{nPeaks}{number of peaks}
}
\value{
dataframe of mean values for each peak number in the input vendor data
}
\description{
average_times: function that finds the averages of each peak time for vendor data (called in find_expected_times())
}
\examples{
Usage Example
peak15avg.df<-average_times(peaks15.list,15)
}
