% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{bivar}
\alias{bivar}
\title{Calculate bivariate correlation}
\usage{
bivar(X, Y)
}
\arguments{
\item{X}{data.frame}

\item{Y}{data.frame}
}
\description{
Calculate the p-values of Pearson's correlation between any
combination of variables from data.frame X and Y
}
