% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{build_expected_times}
\alias{build_expected_times}
\title{build_expected_times: function to build expected times data for reference and sample peaks}
\usage{
build_expected_times(ref.ind, sample.ind, avgs.df)
}
\arguments{
\item{ref.ind}{numeric vector containing the indices of reference peaks in avgs.df}

\item{sample.ind}{numeric vector containing the indices of sample peaks in avgs.df}

\item{avgs.df}{dataframe of average values for peaks times}
}
\value{
list of 2 elements: first for reference peaks expected times and the second for sample peaks
}
\description{
build_expected_times: function to build expected times data for reference and sample peaks
}
\examples{
Usage Example
refPk.vec<-c(1,2,4,5,16)
samplePk.vec<-c(3,6,7:15)
pkAvg.df<-average_times(peaks15.list,15)
expTimes.list<-build_expected_times(ref.ind,sample.ind,avgs.df)
}
