% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{cha}
\alias{cha}
\title{cha: function to find the area of a convex hull}
\usage{
cha(x, y)
}
\arguments{
\item{x}{dataframe of x values}

\item{y}{dataframe of y values}
}
\value{
area of the polygon
}
\description{
cha: function to find the area of a convex hull
}
\examples{
Usage example
cha(peak1t,peak1v)
}
