% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{check_picked_and_vendor_peaks}
\alias{check_picked_and_vendor_peaks}
\title{check_picked_and_vendor_peaks: function that checks peak times and numbers from vendor table against picked peaks}
\usage{
check_picked_and_vendor_peaks(
  vend.df,
  int.mat,
  time.vec,
  z.thresh = 5,
  graphLab,
  timeThresh = 2,
  expPeaks = c(15, 16)
)
}
\arguments{
\item{vend.df}{dataframe of vendor data}

\item{int.mat}{1-col matrix of intensity values}

\item{time.vec}{numeric vector of time values}

\item{z.thresh}{z-value threshold for ThresholdingAlgo}

\item{graphLab}{label for graph of picked times}

\item{timeThresh}{threshold for time differences between picked and vendor peaks}

\item{expPeaks}{vector containg two values for expected numbers of peaks}
}
\value{
list of 3 elements: [\link{1}] dataframe indicating whether the number of peaks in
vendor table match number picked; [\link{2}] dataframe containing number of vendor peaks
and picked peaks; [\link{3}] dataframe of peak times for vendor and picked peaks
}
\description{
check_picked_and_vendor_peaks: function that checks peak times and numbers from vendor table against picked peaks
}
\examples{
Usage Example
int.mat<-matrix(v44,ncol=1) # for ThresholdingAlgo
time.vec<-time.s
graphLab<-"170525_NaHCO3 L + NaCl L"
numPeakCheck<-check_picked_and_vendor_peaks(vend1.df,int.mat,time.s,5,graphLab,2)
}
