% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{copy_Identifier_1_files}
\alias{copy_Identifier_1_files}
\title{move_Identifier_1_files: function that copies all files with a specified identifier into a folder labeled with the Identifier_1 value}
\usage{
copy_Identifier_1_files(identifier_1_files, path, identifier_1)
}
\arguments{
\item{identifier_1_files}{files that contain the desired identifier}

\item{path}{location of the file to be copied}

\item{identifier_1}{the identifier_1 of the files to be copied}
}
\description{
move_Identifier_1_files: function that copies all files with a specified identifier into a folder labeled with the Identifier_1 value
}
\examples{
Usage example
copy_Identifier_1_files(identifier_1_d_files,pathc,identifier_1_d)
}
