% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlms.R
\name{covar}
\alias{covar}
\title{Calculate covariants}
\usage{
covar(X, design, Z, bivars)
}
\arguments{
\item{X}{data.frame}

\item{design}{matrix}

\item{Z}{data.frame}

\item{bivars}{matrix returned by \code{\link{bivar}}}
}
\description{
Calculate the p-values of the variable Z as a covariate in the
complete model. Only variables that are significantly correlated with the x
variabel in X is tested.
}
\keyword{internal}
