% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{d18O_samples_list}
\alias{d18O_samples_list}
\title{d18O_samples_list: function to get sample peak and d18O data for the specified files}
\usage{
d18O_samples_list(standardFileNames, expPk.num = 5, exp.df)
}
\arguments{
\item{standardFilesNames}{character vector of filenames containing standards (of same type)}
}
\value{
list of dataframes containing sample peaks and d18O data
}
\description{
d18O_samples_list: function to get sample peak and d18O data for the specified files
}
\examples{
Usage example
LWstandFiles<-find_files_by_analysis_num(LWiso,LWstandardAnalysisNum)
LW_d18O.list<-d18O_samples_list(LWstandardFileNames)
}
