% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlms.R
\name{export.summary.mlms}
\alias{export.summary.mlms}
\title{export summary mlms to excel}
\usage{
\method{export}{summary.mlms}(x, file)
}
\arguments{
\item{x}{summary.mlms returned by \code{\link{summary.mlms}}}

\item{file}{character. The file path to export}
}
\description{
export summary mlms to excel files
}
\examples{
data(growth)
X = growth[,1:3]
design = model.matrix(~treatment, data = growth)
Z = growth[,5:10]
fit = fit_mlms(X, design, Z, coef = "treatmentLNS")
res = summary(fit)
export(res, "growth.xlsx")
}
