% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{extract_rintensity_all_tsfeatures}
\alias{extract_rintensity_all_tsfeatures}
\title{extract_rintensity_all_tsfeatures: extract time series features from Intensity_All using tsfeatures}
\usage{
extract_rintensity_all_tsfeatures(rintensity_all.num)
}
\arguments{
\item{rintensity_all.num}{numeric vector containing the rIntensity_All data}
}
\value{
dataframe containing the extracted tsfeatures of the Intensity_All data
}
\description{
extract_rintensity_all_tsfeatures: extract time series features from Intensity_All using tsfeatures
}
\examples{
Usage Example
feat<-extract_intensity_all_tsfeatures(int_all_num)
}
