% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{extract_tsfeatures}
\alias{extract_tsfeatures}
\title{extract_tsfeatures: extract time series features from intensity data using tsfeatures}
\usage{
extract_tsfeatures(intensity.data)
}
\arguments{
\item{rintensity_all.num}{numeric vector containing the rIntensity_All data}
}
\value{
dataframe containing the extracted tsfeatures of the rIntensity_All data
}
\description{
extract_tsfeatures: extract time series features from intensity data using tsfeatures
}
\examples{
Usage Example
extract_rintensity_all_tsfeatures(int_all_num)
}
