% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{file_move}
\alias{file_move}
\title{file_move: function to move a file from one directory to another, and to create that directory if it does not exist}
\usage{
file_move(from, to)
}
\arguments{
\item{from}{the original file directory}

\item{to}{the new file directory}
}
\description{
file_move: function to move a file from one directory to another, and to create that directory if it does not exist
}
\examples{
Usage example
file_move(orig_dir,new_dir)
}
