% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{filterSpecSignals}
\alias{filterSpecSignals}
\title{filterSpecSignals: function that filters out possible false positives in peak detection}
\usage{
filterSpecSignals(threshAlgoList)
}
\arguments{
\item{threshAlgoList}{list returned by ThresholdingAlgo()}
}
\value{
list that has been filtered for possible false positive signals (3 consecutive 1s or less in signal matrix)
}
\description{
filterSpecSignals: function that filters out possible false positives in peak detection
}
\examples{
Usage Example
filterSpecSignals(ThreshAlgoList)
}
