% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{find_expected_times}
\alias{find_expected_times}
\title{find_expected_times: function to find average values for start, Rt and end times}
\usage{
find_expected_times(peaks_times.list, expNumPks = c(15, 16))
}
\arguments{
\item{peaks_times.list}{a list of dataframes of peak numbers and times data}
}
\value{
if data for both peak numbers present, a list of dataframes of expected times for data
with both peak numbers, if data of just one of the expected peak numbers is found, it returns a dataframe of expected times
}
\description{
find_expected_times: function to find average values for start, Rt and end times
}
\examples{
Usage Example
pT.df<-find_expected_times(peaksTimes.list)
}
