% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{find_files_by_analysis_num}
\alias{find_files_by_analysis_num}
\title{find_files_by_analysis_num: function that gets filenames of files with specified analysis numbers}
\usage{
find_files_by_analysis_num(standards.Iso, analysisNum.vec)
}
\arguments{
\item{standards.Iso}{continuous flow data for standards files (of the same standard)}

\item{analysisNum.vec:}{numeric vector of analysis numbers identifying which standard files to pull}
}
\value{
character vector of file names containing the specified analyses
}
\description{
find_files_by_analysis_num: function that gets filenames of files with specified analysis numbers
}
\examples{
Usage Example
LWstandardFileNames<-find_files_by_analysis_num(LWiso,LWstandardAnalysisNum)
}
