% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_mlms}
\alias{fit_mlms}
\title{fit mixed linear model selection}
\usage{
fit_mlms(X, design, Z, coef, alternative = "two.sided")
}
\arguments{
\item{X}{data.frame}

\item{design}{matrix. Output from \code{\link{model.matrix}}. Number of rows
must equal to that of X.}

\item{Z}{data.frame. Number of rows must equal to that of X.}

\item{coef}{character. The target coefficient to test.}

\item{alternative}{character. The length must be either 1 or the same as the
number of columns in X.}
}
\value{
fit_mlms returns an object of class "mlms". The function
\code{summary} can be further used to organize the result.

An object of class "mlms" is a list containing the following components:

\describe{
  \item{bivars}{The p-value matrix for bivariate correlation between any
  variable from \code{X} and \code{Z}.}
  \item{covars}{The p-value matrix for covariates in the complete model.}
  \item{df.residuel}{Residual degree of freedom}
  \item{confint}{95\% confidence internal}
  \item{unadjusted}{The coefficients of unadjusted model}
  \item{adjusted}{The coefficients of adjusted model}
}
}
\description{
This function fits a linear model regression between any
variable in \code{data.frame} X and the design matrix, with the
\code{data.frame} Z as covariates.
}
\details{
Each potential covariate in Z is frist tested whether it is
correlated with any variable in X. The result of this step is a \code{matrix}
of p values in the slot \code{bivar}. For any variable in X, only the
covariants in Z with a p-value < 0.1 is used to adjust the variance in the
complete model. Next, the covariants are tested whether they are significant
in the complete model. The result of this step is in the slot \code{covar}.
Finally, the complete model is fit to test the effect of the target
coefficient. The result is in the slot \code{fits}
}
\examples{
data(growth)
X = growth[,1:3]
design = model.matrix(~treatment, data = growth)
Z = growth[,5:10]
fit_mlms(X, design, Z, coef = "treatmentLNS")
}
