% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{generic_raw_plot}
\alias{generic_raw_plot}
\title{generic_raw_plot: function to plot raw data using generic plot}
\usage{
generic_raw_plot(raw.df, title)
}
\arguments{
\item{raw.df}{full dataframe of raw data}

\item{title}{title of the plot}
}
\description{
generic_raw_plot: function to plot raw data using generic plot
}
\examples{
Usage Example
generic_raw_plot(rawDat.df,"170525_NaHCO3 L + NaCl L_.dxf")
}
