% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{get_all_filenames}
\alias{get_all_filenames}
\title{get_all_filenames: function to get all file names from a directory of .dxf files}
\usage{
get_all_filenames(path)
}
\arguments{
\item{path}{path to the directory of .dxf files}
}
\value{
vector of filenames for all .dxf files in the specified directory
}
\description{
get_all_filenames: function to get all file names from a directory of .dxf files
}
\examples{
Usage example
get_all_filenames("~/path_to_files")
}
