% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{get_all_peak_areas}
\alias{get_all_peak_areas}
\title{get_all_peak_areas: function to get all peak areas in an experiment}
\usage{
get_all_peak_areas(start.vec, end.vec, time.vec, int.vec)
}
\arguments{
\item{start.vec}{numeric vector containing all peak start times}

\item{end.vec}{numeric vector containing all peak end times}

\item{time.vec}{numeric vector containing all the raw time.s data}

\item{int.vec}{numeric vector continaing all the raw intentisty data (i.e. v44.mv, etc,)}
}
\value{
a vector containing the areas of each peak
}
\description{
get_all_peak_areas: function to get all peak areas in an experiment
}
\examples{
Usage example
get_all_peak_areas(start.v1,end.v1,allt.s,allv44)
}
