% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{get_identifier_1_files}
\alias{get_identifier_1_files}
\title{get_identifier_1_files: function to get filenames whose Identifier_1 data matches the one specified}
\usage{
get_identifier_1_files(filenames, identifier_1, cores = 2)
}
\arguments{
\item{filenames}{vector of filenames}

\item{identifier_1}{the name of the desired Identifier_1}

\item{cores}{number of cores to use for the grepl function to search through the filenames vector}
}
\value{
dataframe of filenames with the specified Identifier_1
}
\description{
get_identifier_1_files: function to get filenames whose Identifier_1 data matches the one specified
}
\examples{
Usage example
identifier_files<-get_identifier_1_files(my_filenames,my_identifier_1)
}
