% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{get_raw_df}
\alias{get_raw_df}
\title{get_raw_df: get the raw data from .dxf files as a dataframe}
\usage{
get_raw_df(files)
}
\arguments{
\item{files}{vector containing character strings of .dxf filenames}
}
\value{
dataframe containing all raw data in the .dxf files
}
\description{
get_raw_df: get the raw data from .dxf files as a dataframe
}
\examples{
Usage example
#' data_files<-c("170525_NaHCO3 L + NaCl L_.dxf","170525_NaHCO3 L + NaCl U_.dxf","170525_NaHCO3 L_.dxf","170525_NaHCO3 U + NaCl L_.dxf",
"170525_NaHCO3 U + NaCl U_.dxf","170525_NaHCO3 U_.dxf")
get_raw_df(data_files)
}
