% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{get_reference_values_no_ratio}
\alias{get_reference_values_no_ratio}
\title{get_reference_values_no_ratio: get isotopic reference values as a dataframe}
\usage{
get_reference_values_no_ratio(files)
}
\arguments{
\item{files}{vector containing character strings of .dxf file names}
}
\value{
dataframe of reference values in the .dxf files
}
\description{
get_reference_values_no_ratio: get isotopic reference values as a dataframe
}
\examples{
Usage example
data_files<-c("170525_NaHCO3 L + NaCl L_.dxf","170525_NaHCO3 L + NaCl U_.dxf","170525_NaHCO3 L_.dxf","170525_NaHCO3 U + NaCl L_.dxf",
"170525_NaHCO3 U + NaCl U_.dxf","170525_NaHCO3 U_.dxf")
stand_no_ratio<-get_reference_values_no_ratio(data_files)
}
