% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{get_unique_identifiers}
\alias{get_unique_identifiers}
\title{get_unique_identifiers: Function to find all of the different Identifier_1 labels in a directory of .dxf files}
\usage{
get_unique_identifiers(path, filenames)
}
\arguments{
\item{all_Identifier_1_labels}{vector of the Identifier_1 labels of each .dxf file in the directory}
}
\value{
Usage example
get_unique_identifiers(double_component_Identifier_1s)
}
\description{
get_unique_identifiers: Function to find all of the different Identifier_1 labels in a directory of .dxf files
}
