% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{group_picked_times}
\alias{group_picked_times}
\title{group_picked_times: function that organizes the times picked from the PeakTimes function
into single average times for peaks}
\usage{
group_picked_times(timesPicked)
}
\arguments{
\item{timesPicked}{the times returned in a list from ThresholdingAlgo}
}
\value{
the average times from each "group" (groups returned from ThresholdingAlgo$signals that needs to be separated)
}
\description{
group_picked_times: function that organizes the times picked from the PeakTimes function
into single average times for peaks
}
\examples{
Usage Example
group_picked_times(timesPicked.list)
}
