% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{groupPickedTimes}
\alias{groupPickedTimes}
\title{groupPickedTimes: function that organizes the times picked from the PeakTimes function
into single average times for peaks}
\usage{
groupPickedTimes(timesPicked)
}
\arguments{
\item{timesPicked}{the times returned in a list from ThresholdingAlgo}
}
\value{
the average times from each "group" (groups returned from ThresholdingAlgo$signals that needs to be separated)
}
\description{
groupPickedTimes: function that organizes the times picked from the PeakTimes function
into single average times for peaks
}
\examples{
Usage Example
groupPickedTimes(timesPicked.list)
}
