% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{identifier_1_files}
\alias{identifier_1_files}
\title{identifier_1_files: function to get filenames whose Identifier_1 data matches the one specified}
\usage{
identifier_1_files(files, identifier_1, cores = 2)
}
\arguments{
\item{files}{vector of file names}

\item{identifier_1}{the name of the desired Identifier_1}

\item{cores}{number of cores to use for the grepl function to search through the files vector (default=2)}
}
\value{
dataframe of file names with the specified Identifier_1
}
\description{
identifier_1_files: function to get filenames whose Identifier_1 data matches the one specified
}
\examples{
Usage example
identifier_1_files(my_filenames,my_identifier_1)
}
