% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{intensity_similarity_check}
\alias{intensity_similarity_check}
\title{intensity_similarity_check: function that checks specified peaks for similarity in intensity values}
\usage{
intensity_similarity_check(
  vendAmpl,
  amplName,
  peakNr.vec,
  relDiff.thresh = 0.1
)
}
\arguments{
\item{vendAmpl}{vector of amplitudes (as from vendor table)}

\item{amplName}{label of the amplitude (as from vendor table)}

\item{peakNr.vec}{vector of peak numbers of interest}

\item{relDiff.thresh}{threshold for relative difference between intensities}
}
\value{
list that contains TRUE/FALSE if intensities are within specified threshold and a dataframe of the peak numbers and intensities analyzed
}
\description{
intensity_similarity_check: function that checks specified peaks for similarity in intensity values
}
\examples{
Usage Example
intensity_similarity_check(vend.df$Ampl_44,"Ampl_44",c(1,2,4,5,16),0.1)
}
