% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dl_button.R
\name{make_dl_button}
\alias{make_dl_button}
\title{Create HTML Download Button}
\usage{
make_dl_button(table_id, ver = utils::packageVersion("mlms"))
}
\arguments{
\item{table_id}{'character' string.
Element identifier for the table widget.}

\item{ver}{'package_version' or 'character' string.
Version identifier inserted into the CSV file name.}
}
\value{
Returns an object of class 'shiny.tag'.
}
\description{
Create a download button that exports the selected table data
to a comma-separated values (CSV) file when clicked.
}
\examples{
html <- make_dl_button("test")
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{internal}
