% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{make_dummy_variables}
\alias{make_dummy_variables}
\title{make dummy variables}
\usage{
make_dummy_variables(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame
}
\description{
convert any categorical variables into dummy variables in a data
frame
}
\examples{
df = make_dummy_variables(iris)
}
