% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlms.R
\docType{package}
\name{mlms-package}
\alias{mlms}
\alias{mlms-package}
\title{mlms: Multilevel Monitoring System Data for Wells in the USGS INL Aquifer Monitoring Network}
\description{
Analysis-ready datasets detailing the Multilevel Monitoring System (MLMS) wells within the U.S. Geological Survey's (USGS) aquifer-monitoring network at the Idaho National Laboratory (INL) in Idaho, and the data collected within these wells. Supported by the U.S. Department of Energy (DOE), the USGS collected discrete measurements of hydraulic head at various depths from wells in the eastern Snake River Plain (ESRP) aquifer over several years. These measurements were derived from data on fluid pressure, fluid temperature, and atmospheric pressure. Each well was equipped with an MLMS, which included valved measurement ports, packer bladders, casing segments, and couplers. The MLMS facilitated monitoring at multiple hydraulically isolated depth intervals, reaching significant depths below the land surface. Additionally, groundwater samples were collected from these wells over multiple years and analyzed for various chemical and physical parameters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://code.usgs.gov/inl/mlms}
  \item Report bugs at \url{https://code.usgs.gov/inl/mlms/-/issues}
}

}
\author{
\strong{Maintainer}: Jason C. Fisher \email{jfisher@usgs.gov} (\href{https://orcid.org/0000-0001-9032-8912}{ORCID})

Authors:
\itemize{
  \item Brian V. Twining \email{btwining@usgs.gov} (\href{https://orcid.org/0000-0003-1321-4721}{ORCID})
}

}
\keyword{internal}
