% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{peak_area}
\alias{peak_area}
\title{peak_area: Function to calculate the area under a peak}
\usage{
peak_area(start.t, end.t, time.vec, int.vec)
}
\arguments{
\item{start.t}{start time of the peak}

\item{end.t}{end time of the peak}

\item{time.vec}{the raw time.s data}

\item{int.vec}{the raw intensity data for a specified mass}
}
\value{
the area under the specified peak
}
\description{
peak_area: Function to calculate the area under a peak
}
\examples{
Usage example
peak_area(start1,end1,allt.s,allv44)
}
