% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{peak_times_check}
\alias{peak_times_check}
\title{peak_times_check: function that checks whether peaks occur at the expected times}
\usage{
peak_times_check(
  vend.df,
  expectedPeak.num,
  diff.t = 10,
  expectedStart,
  expectedRt,
  expectedEnd
)
}
\arguments{
\item{vend.df}{dataframe of vendor table info}

\item{expectedPeak.num}{expected number of reference peaks (default=5)}

\item{diff.t}{acceptable time difference between expected peak times and actual peak times (default=5s)}

\item{expectedStart}{vector of expected peak Start times (default for CO2)}

\item{expectedRt}{vector of expected peak Rt (default for CO2)}

\item{expectedEnd}{vector of expected peak End times (default for CO2)}
}
\value{
list containing boolean values indicating whether the expected number of reference peaks was found at the expected times and a dataframe
containing the peak numbers and times
}
\description{
peak_times_check: function that checks whether peaks occur at the expected times
}
\examples{
Usage Example
expTimes.list<-build_expected_times(refInd.vec,sampInd.vec,avgs.df)
peak_times_check(vendor.list[[10]],expectedStart=expTimes.list[[1]]$Expected_Start,expectedRt=expTimes.list[[1]]$Expected_Rt,expectedEnd=expTimes.list[[1]]$Expected_End) # using default values
}
