% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{peak_times_pick}
\alias{peak_times_pick}
\title{peak_times_pick: function that gives the approximate peak start times}
\usage{
peak_times_pick(Int.mat, time.interval, z.thresh, reactionLabel)
}
\arguments{
\item{Int.mat}{matrix of intensity values within the desired interval}

\item{time.interval}{time values in the desired interval}

\item{z.thresh}{z-value threshold for ThresholdAlgo}

\item{reactionLabel}{label for the reaction (for plot generated showing peak start times)}
}
\value{
vector of average peak start times as found by ThresholdingAlgo and groupPickedTimes
}
\description{
peak_times_pick: function that gives the approximate peak start times
}
\examples{
Usage Example
peak_times_pick(v44Int.mat,interval.times)
}
