% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{peak_vend_times}
\alias{peak_vend_times}
\title{peak_vend_times: function that returns the peak numbers and times for a given vendor table}
\usage{
peak_vend_times(vend.df)
}
\arguments{
\item{vend.df}{dataframe of vendor data}
}
\value{
dataframe containing peak numbers, start, retention times, and end times
}
\description{
peak_vend_times: function that returns the peak numbers and times for a given vendor table
}
\examples{
Usage example
pkTimes.df<-peak_vend_times(v1.df)
}
