% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{peakTimes}
\alias{peakTimes}
\title{peakTimes: function}
\usage{
peakTimes(Int.mat, time.interval)
}
\arguments{
\item{Int.mat}{matrix of intensity values within the desired interval}

\item{time.interval}{time values in the desired interval}
}
\value{
vector of average peak start times as found by ThresholdingAlgo and groupPickedTimes
}
\description{
peakTimes: function
}
\examples{
Usage Example
PeakTimes(v44Int.mat,interval.times)
}
