% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{plot_all_peaks}
\alias{plot_all_peaks}
\title{plot_all_peaks: Function that plots every peak in an experiment in a separate graph}
\usage{
plot_all_peaks(start.v, end.v, t.s, v.mV, v.name)
}
\arguments{
\item{start.v}{numeric vector of start times for all the peaks in the experiment}

\item{end.v}{numeric vector of end times for all the peaks in the experiment}

\item{t.s}{numeric vector time.s from raw data}

\item{v.mV}{numeric vector of raw voltages (i.e. v44.mV, v45.mV, v46.mV, etc.)}

\item{v.name}{the name of raw voltage column}
}
\description{
plot_all_peaks: Function that plots every peak in an experiment in a separate graph
}
\examples{
Usage example
plot_all_peaks(start.v1,end.v1,allt.s,allv44,"v44.mV")
}
