% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{plot_individual_peak}
\alias{plot_individual_peak}
\title{plot_individual_peak: function to plot an individual peak in an experiment}
\usage{
plot_individual_peak(start.t, end.t, time.vec, int.vec, peak_num, v.mv)
}
\arguments{
\item{start.t}{start time of the peak}

\item{end.t}{end time of the peak}

\item{time.vec}{vector containing all time.s raw data}

\item{int.vec}{vectoring containing the raw intensity data for a specified mass (i.e. v44.mV, etc.)}

\item{peak_num}{the peak number}

\item{v.mv}{vector containing the raw mV intensity data}
}
\description{
plot_individual_peak: function to plot an individual peak in an experiment
}
\examples{
Usage example
plot_individual_peak(start1,end1,allt.s,allv44,"1","v44.mV")
}
