% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{qc_CO2_num_peaks}
\alias{qc_CO2_num_peaks}
\title{qc_num_peaks: Function that checks for the correct number of peaks}
\usage{
qc_CO2_num_peaks(vend.df)
}
\arguments{
\item{vend.df}{dataframe of vendor data that contains Peak_Nr}
}
\value{
Boolean that indicates whether the expected number of peaks is present
}
\description{
qc_num_peaks: Function that checks for the correct number of peaks
}
\examples{
Usage example
("Good" CO2 peak will have all 3 masses stacked nicely on top of each other)
qc_num_peaks(vend1.df)
}
