% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{qc_num_vendPeaks}
\alias{qc_num_vendPeaks}
\title{qc_num_vendPeaks: Function that checks for the correct number of peaks}
\usage{
qc_num_vendPeaks(vend.df, expectedPeakNum1 = 15, expectedPeakNum2 = 16)
}
\arguments{
\item{vend.df}{dataframe of vendor data that contains Peak_Nr}

\item{expectedPeakNum1}{first value of expected peak numbers (i.e. if a peak can be missing - set both to the same number if no variation in peak numbers is expected)}

\item{expectedPeakNum2}{second value of expected peak numbers}
}
\value{
Boolean that indicates whether the expected number of peaks is present
}
\description{
qc_num_vendPeaks: Function that checks for the correct number of peaks
}
\examples{
Usage example
qc_num_vendPeaks(vend1.df,15,16)
}
