% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{qc_peaks_vend}
\alias{qc_peaks_vend}
\title{qc_peaks_vend: Function to check for peak presence}
\usage{
qc_peaks_vend(start.times)
}
\arguments{
\item{start.times}{vector of start times for each peak in the data}
}
\value{
Boolean that indicates whether peaks are present
}
\description{
qc_peaks_vend: Function to check for peak presence
}
\examples{
Usage example
qc_peaks_vend(start_times.vec)
}
