% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{qc_samplePeaks}
\alias{qc_samplePeaks}
\title{qc_samplePeaks: function that tests whether the expected number of sample peaks are present and whether they occur at the expected times}
\usage{
qc_samplePeaks(
  intensity.mat,
  time.interval,
  expectedNum.samplePeaks,
  expectedStartSamples,
  samplePeakTimes
)
}
\arguments{
\item{intensity.mat}{matrix of intensity values for the desired interval}

\item{time.interval}{numeric vector of time values for the desired interval}

\item{expectedNum.samplePeaks}{number of expected sample peaks}

\item{expectedStartSamples}{numeric vector of expected start times}

\item{samplePeakTimes}{numeric vector of times that sample peaks occur, such as that returned by PeakTimes()}
}
\value{
Boolean dataframe with columns "Expected_Number_Sample_Peaks" and "Expected_Start_Times" which indicate whether the expected number of sample peaks and expected start times are found
}
\description{
qc_samplePeaks: function that tests whether the expected number of sample peaks are present and whether they occur at the expected times
}
\examples{
Usage Example
qc_samplePeaks(intensity.mat=v44Int.mat,time.interval=above_below.t,expectedNum.samplePeaks=9,
                expectedStartSamples=start.v1[7:15],samplePeakTimes=peakTimes)
}
