% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sensors_json.R
\name{read_sensors_json}
\alias{read_sensors_json}
\title{Read Sensor Data in JSON Format}
\usage{
read_sensors_json(path)
}
\arguments{
\item{path}{'character' string.
Path to the JSON file to read.}
}
\value{
A list of data frame components.
See \code{\link{sensors}} and \code{\link{calibrations}} datasets for example output.
}
\description{
Read electronic sensor data in a JSON format.
}
\examples{
l <- system.file("extdata/ex-sensors.json", package = "mlms") |>
  read_sensors_json()
str(l, max.level = 1)
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
