% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{read_summ}
\alias{read_summ}
\title{read_summ: read and print a summary of mass spec data from a .dxf file}
\usage{
read_summ(filename)
}
\arguments{
\item{filename}{character string of the name of the .dxf file of data}
}
\value{
summary table of file contents
}
\description{
read_summ: read and print a summary of mass spec data from a .dxf file
}
\examples{
Usage example
}
