% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{read_summary}
\alias{read_summary}
\title{read_summary: read and print a summary of mass spec data from a .dxf file}
\usage{
read_summary(filename)
}
\arguments{
\item{filename}{character string of the name of the .dxf file of data}
}
\value{
summary table of file contents
}
\description{
read_summary: read and print a summary of mass spec data from a .dxf file
}
\examples{
Usage example
read_summ("170525_NaHCO3 L + NaCl L_.dxf")
}
