% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{reference_times_check}
\alias{reference_times_check}
\title{reference_times_check: function to determine if the expected number and start,end rt times of reference peaks are present}
\usage{
reference_times_check(
  vend.df,
  expectedPeak.num,
  diff.t = 10,
  expectedStart = c(27.17, 67.0193, 166.5653, 206.492, 843.3227),
  expectedRt = c(47.3888, 87.184, 186.7608, 226.5096, 863.5029),
  expectedEnd = c(50.4929, 90.3422, 189.8572, 229.6987, 866.6069)
)
}
\arguments{
\item{vend.df}{dataframe of vendor data}

\item{expectedPeak.num}{number of expected peaks}

\item{diff.t}{acceptable difference between peak times and expected times}

\item{expectedStart}{vector of expected start times}

\item{expectedRt}{vector of expected retention times}

\item{expectedEnd}{vector of expected end times}
}
\value{
list of 3 elements: [\link{1}] boolean-if peaks are at expected times; [\link{2}] peak numer and times vendor data
[\link{3}] full vendor data for reference peaks
}
\description{
reference_times_check: function to determine if the expected number and start,end rt times of reference peaks are present
}
\examples{
Usage example
refTC<-reference_times_check(vi.list[[1]])
}
