% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{reference_values_ratio}
\alias{reference_values_ratio}
\title{reference_values_ratio: get isotopic reference values as a dataframe}
\usage{
reference_values_ratio(files)
}
\arguments{
\item{files}{vector containing character strings of .dxf file names}
}
\value{
dataframe of reference values in the .dxf files
}
\description{
reference_values_ratio: get isotopic reference values as a dataframe
}
\examples{
Usage example
reference_values_ratio(data_files)
}
