% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{remove_276}
\alias{remove_276}
\title{remove_276: function to remove the (1st sample) peak at 276 which often has some contamination for CO2}
\usage{
remove_276(vend.df)
}
\arguments{
\item{vend.df}{dataframe of full vendor table data}
}
\value{
new vend.df without the 276 peak data
}
\description{
remove_276: function to remove the (1st sample) peak at 276 which often has some contamination for CO2
}
\examples{
Usage Example
remove_276(vend1.df)
}
