% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{sample_peaks}
\alias{sample_peaks}
\title{sample_peaks: function to grab the vendor info for the peaks at expected sample peak times}
\usage{
sample_peaks(vend.df, start.sample = 326, stop.sample = 825)
}
\arguments{
\item{vend.df}{vendor info dataframe of all peaks}

\item{start.sample}{approximate expected start time for sample peaks  (default 326s)}

\item{stop.sample}{approximate expected stop time for sample peaks (default 825s)}
}
\value{
vendor table for just the (sample) peaks from the given start and stop times
}
\description{
sample_peaks: function to grab the vendor info for the peaks at expected sample peak times
}
\examples{
Usage Example
sample_peaks(no276Vend1.df)
}
