% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{sample_peaks_process}
\alias{sample_peaks_process}
\title{sample_peaks_process: function to remove flush peak if present and 1st sample peak}
\usage{
sample_peaks_process(
  refTimesOutput,
  vend.df,
  flushExpT = 135,
  flushTint = 10,
  firstSampExpT = 275,
  firstSampTint = 10
)
}
\arguments{
\item{refTimesOutput}{output from reference_times_check (a list)}

\item{vend.df}{dataframe of vendor data}

\item{flushExpT}{expected Rt of the flush peak}

\item{flushTint}{acceptable difference in Rt between flush peak Rt and expected time}

\item{firstSampTint}{acceptable difference in Rt between first sample peak Rt and expected time}

\item{firstSampleT}{expected Rt of the first sample peak}
}
\value{
dataframe of processed sample peak vendor data
}
\description{
sample_peaks_process: function to remove flush peak if present and 1st sample peak
}
\examples{
Usage Example
samplePeaksVend.df<-sample_peaks_process(refTC,vend1.df)
}
