% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{select_file_info}
\alias{select_file_info}
\title{select_file_info:}
\usage{
select_file_info(filename)
}
\arguments{
\item{filename}{character string of the name of the .dxf file of data}
}
\value{
file information - Name, Analysis, Peak Center, H3 Factor, Date & Time
}
\description{
select_file_info:
}
