% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{separate_by_analysis_num}
\alias{separate_by_analysis_num}
\title{separate_by_analysis_num: function to separate out data by analysis number}
\usage{
separate_by_analysis_num(vend.df)
}
\arguments{
\item{vend.df}{dataframe of vendor data for multiple experiments}
}
\value{
list of vendor data separated out by analysis number
}
\description{
separate_by_analysis_num: function to separate out data by analysis number
}
\examples{
Usage Example
sep_analyses<-separate_by_analysis_num(CO2_1p)
}
