\name{simulate_mlHMM}
\alias{simulate_mlHMM}
\title{Simulate data under the Multilevel Hidden Markov Model with a multivariate normal conditional distribution}
\usage{
simulate_mlHMM()
}
\description{
Simulate data under the Multilevel Hidden Markov Model with a multivariate normal conditional distribution. The function allows to simulate random effects in all model parameters. See details.
}
\examples{
data <- simulate_mlHMM()
simulate_mlHMM(data)
}
