\name{simulate_mlMSVAR}
\alias{simulate_mlMSVAR}
\title{Simulate data under the Multilevel Markov switching VAR Model}
\usage{
simulate_mlMSVAR()
}
\description{
Simulate data under the multilevel Markov switching vector autoregressive model.
}
\examples{
data <- simulate_mlMSVAR()
simulate_mlMSVAR(data)
}
