% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{standardInputProc}
\alias{standardInputProc}
\title{standardInputProc: function that gets sample peak and d18O/16O data for internal standards quality check}
\usage{
standardInputProc(path, standardAnalysisNums.vec)
}
\arguments{
\item{path}{path to the .dxf internal standards experiment files}

\item{standardAnalysisNums.vec}{numeric vector of analysis numbers}
}
\value{
list of dataframes containing d18O/16O vendor data for each file
}
\description{
standardInputProc: function that gets sample peak and d18O/16O data for internal standards quality check
}
\examples{
Usage Example
stand_d18O.list<-standardInputProc(standPath,standAnalysisNum)
}
