% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{standards_input_list}
\alias{standards_input_list}
\title{standards_input_list: function to create input list for internal standards quality check}
\usage{
standards_input_list(standardPaths.vec, standardAnalysisNums.df, exp.df)
}
\arguments{
\item{standardPaths.vec}{character vector containing the paths to the .dxf files for internal standards experiments}

\item{standardAnalysisNums.df}{dataframe containing the analysis numbers for internal standards experiments and colnames are the standard names}
}
\value{
list of sample peak data for each analysis
}
\description{
standards_input_list: function to create input list for internal standards quality check
}
\examples{
Usage Example
standAnalysisNums.df<-read.table("standards_Analysis_Nums.txt")
standPaths.vec<-c(L1path,H1path,LWpath)
allStand_d18O.list<-standards_input_list(standPaths.vec,standAnalysisNums.df)
}
