% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlms.R
\name{summary.mlms}
\alias{summary.mlms}
\title{Summarizing mlms result}
\usage{
\method{summary}{mlms}(x, ...)
}
\arguments{
\item{x}{an object of class "mlms", usually, a result of a call to \code{
link{fit_mlms}}}
}
\value{
The function summary.mlms returns a list of result organized from a given
mlms object.
\describe{
  \item{bivars}{The p-value matrix for bivariate correlation between any
  variable from \code{X} and \code{Z}.}
  \item{covars}{The p-value matrix for covariates in the complete model.}
  \item{models}{Combinded model coefficients and statistics for both
  unadjusted and adjusted for each x variable}
}
}
\description{
summary method for class "mlms"
}
\examples{
data(growth)
X = growth[,1:3]
design = model.matrix(~treatment, data = growth)
Z = growth[,5:10]
fit = fit_mlms(X, design, Z, coef = "treatmentLNS")
res = summary(fit)
}
