% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{ThresholdingAlgo}
\alias{ThresholdingAlgo}
\title{ThresholdingAlgo: function based on dispersion
Code from: Robust peak detection algorithm (using z-scores)
https://stackoverflow.com/questions/22583391/peak-signal-detection-in-realtime-timeseries-data?page=1&tab=votes#tab-top}
\usage{
ThresholdingAlgo(y, lag, threshold, influence)
}
\arguments{
\item{y}{intensity matrix (1 col)}

\item{lag}{lag of the moving window}

\item{threshold}{z-score at which the algorithm signals}

\item{influence}{the influence (between 0 and 1) of new signals on the mean and standard deviation}
}
\value{
list of "signals", "avgFilter", "stdFilter" - signal times (signals) are of the most interest for MS purposes
}
\description{
ThresholdingAlgo: function based on dispersion
Code from: Robust peak detection algorithm (using z-scores)
https://stackoverflow.com/questions/22583391/peak-signal-detection-in-realtime-timeseries-data?page=1&tab=votes#tab-top
}
\examples{
Usage Example
ThresholdingAlgo(v44Int.mat,lag=10,threshold=5,influence=0.5)
}
