% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{unique_identifiers}
\alias{unique_identifiers}
\title{unique_identifiers: Function to find all of the different Identifier_1 labels in a directory of .dxf files}
\usage{
unique_identifiers(filenames)
}
\arguments{
\item{filenames}{vector of filenames from which to search for unique Identifier_1 labels}
}
\value{
vector of each different Identifier_1 in the given files
Usage example
unique_identifiers(double_component_files)
}
\description{
unique_identifiers: Function to find all of the different Identifier_1 labels in a directory of .dxf files
}
