% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{vendor_info}
\alias{vendor_info}
\title{vendor_info: get selected vendor info for a .dxf file:
(Identifier 1, Nr., Start, Rt, End, Intensity All, rIntensity All, Ampl_44, Ampl_45, Ampl_46, d13C/12C, d18O/16O)}
\usage{
vendor_info(file)
}
\arguments{
\item{file}{file name}
}
\value{
dataframe of vendor information
}
\description{
vendor_info: get selected vendor info for a .dxf file:
(Identifier 1, Nr., Start, Rt, End, Intensity All, rIntensity All, Ampl_44, Ampl_45, Ampl_46, d13C/12C, d18O/16O)
}
\examples{
Usage Example
vendor_info(file)
}
