% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{vendor_info_all}
\alias{vendor_info_all}
\title{vendor_info_all: function to get vendor table data from several .dxf files}
\usage{
vendor_info_all(files)
}
\arguments{
\item{files}{vector of filenames}
}
\value{
list that contains vendor table info for each file
}
\description{
vendor_info_all: function to get vendor table data from several .dxf files
}
\examples{
Usage example
vendor_info_all(files.vec)
}
