% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlms.R
\name{fit_mlms}
\alias{fit_mlms}
\title{fit mixed linear model selection}
\usage{
fit_mlms(X, design, Z, coef, alternative = "two.sided")
}
\arguments{
\item{X}{data.frame}

\item{design}{matrix. Output from \code{\link{model.matrix}}. Number of rows
must equal to that of X.}

\item{Z}{data.frame. Number of rows must equal to that of X.}

\item{coef}{character. The target coefficient to test.}

\item{alternative}{character. The length must be either 1 or the same as the
number of columns in X.}
}
\description{
This function fits a linear model regression between any
variable in \code{data.frame} X and the design matrix, with the
\code{data.frame} Z as covariates.
}
\details{
Each potential covariate in Z is frist tested whether it is
correlated with any variable in X. The result of this step is a \code{matrix}
of p values in the slot \code{bivar}. For any variable in X, only the
covariants in Z with a p-value < 0.1 is used to adjust the variance in the
complete model. Next, the covariants are tested whether they are significant
in the complete model. The result of this step is in the slot \code{covar}.
Finally, the complete model is fit to test the effect of the target
coefficient. The result is in the slot \code{fits}
}
\examples{
data(growth)
X = growth[,1:3]
design = model.matrix(~treatment, data = growth)
Z = growth[,5:10]
fit_mlms(X, design, Z, coef = "treatmentLNS")
}
