% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{get_raw_df}
\alias{get_raw_df}
\title{get_raw_df: get the raw data from .dxf files as a dataframe}
\usage{
get_raw_df(files)
}
\arguments{
\item{files}{vector containing character strings of .dxf file names}
}
\value{
dataframe containing all raw data in the .dxf files
}
\description{
get_raw_df: get the raw data from .dxf files as a dataframe
}
\examples{
Usage example
get_raw_df(data_files)
}
