% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{get_resistor_df}
\alias{get_resistor_df}
\title{get_resistor_df: get resistor info for collection of .dxf files as a dataframe}
\usage{
get_resistor_df(files)
}
\arguments{
\item{files}{vector containing character strings of .dxf file names}
}
\value{
dataframe of resistor info in the .dxf files
}
\description{
get_resistor_df: get resistor info for collection of .dxf files as a dataframe
}
\examples{
Usage example
get_resistor_df(data_files)
}
