% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{select_vendor_info}
\alias{select_vendor_info}
\title{select_vendor_info: get specific vendor info with labeled experiment names for a collection of .dxf files
(Identifier 1, Nr., Start, Rt, End, Intensity All, d13C/12C, d18O/16O)}
\usage{
select_vendor_info(files)
}
\arguments{
\item{files}{vector containing character strings of .dxf file names}
}
\value{
dataframe of vendor information with rows labeled with experiment name (Identifier 1)
}
\description{
select_vendor_info: get specific vendor info with labeled experiment names for a collection of .dxf files
(Identifier 1, Nr., Start, Rt, End, Intensity All, d13C/12C, d18O/16O)
}
\examples{
Usage Example
data_files<-c("170525_NaHCO3 L + NaCl L_.dxf","170525_NaHCO3 L + NaCl U_.dxf","170525_NaHCO3 L_.dxf","170525_NaHCO3 U + NaCl L_.dxf",
"170525_NaHCO3 U + NaCl U_.dxf","170525_NaHCO3 U_.dxf")
vend_info<-select_vendor_info(data_files)
}
