% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{sort_by_identifier_1}
\alias{sort_by_identifier_1}
\title{sort_by_identifier_1: Function that sorts all .dxf files in a path into folders labeled with the Identifier_1 labels}
\usage{
sort_by_identifier_1(path)
}
\arguments{
\item{path}{path to files to be sorted}
}
\description{
sort_by_identifier_1: Function that sorts all .dxf files in a path into folders labeled with the Identifier_1 labels
}
\examples{
Usage example
sort_by_identifier_1("~lily/Desktop/Europa MLMS/Europa_Data/Europa double comp")
}
