% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{unique_identifiers}
\alias{unique_identifiers}
\title{unique_identifiers: Function to find all of the different Identifier_1 labels in a directory of .dxf files}
\usage{
unique_identifiers(path, filenames)
}
\arguments{
\item{all_Identifier_1_labels}{vector of the Identifier_1 labels of each .dxf file in the directory}
}
\value{
Usage example
unique_identifiers(double_component_Identifier_1s)
}
\description{
unique_identifiers: Function to find all of the different Identifier_1 labels in a directory of .dxf files
}
