% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLMSfunctions.R
\name{vendor_info}
\alias{vendor_info}
\title{vendor_info: get selected vendor info with labeled experiment names for a collection of .dxf files
(Identifier 1, Nr., Start, Rt, End, Intensity All, rIntensity All, d13C/12C, d18O/16O)}
\usage{
vendor_info(files)
}
\arguments{
\item{files}{vector containing character strings of .dxf file names}
}
\value{
dataframe of vendor information with rows labeled with experiment name (Identifier 1)
}
\description{
vendor_info: get selected vendor info with labeled experiment names for a collection of .dxf files
(Identifier 1, Nr., Start, Rt, End, Intensity All, rIntensity All, d13C/12C, d18O/16O)
}
\examples{
Usage Example
select_vendor_info(data_files)
}
